Functionality to view performance metrics for each test image with optimized weights obtained from each of the 10 training images
Hi guys, so I was able to add the functionality to print Fh and Ffa for each of the 10 testing images obtained for the optimized weights for each of the 10 training images as a table and csv file. So, overall our table will have 100 entries. We can now see which weights optimize performance metrics to the maximum extent.